Added weather datasets
Added a script to retrieve the latitude, longitude and datetime of each record in the municipal issue datasets, and then use it to perform a query on the Open-Meteo Weather API to extract the hourly weather & wind data of that specific area. For each record, the datetime queried includes the hour the report was made, and in a 1 hour interval until 24h before the datetime the report was made. This is done to account for the time lag between the time at which the report was made, and the time at which the issue actually occurred.

NOTE: Open-Meteo has a daily limit of 10k API calls under a free plan.