Update municipal data retrieval logic
Updated the filename to socrata_opendata_fetcher.py to more accurately
reflect it's intended function of fetching via Socrata.

But more importantly, as Weather API's free plan tend to have a low
daily limit of API calls per user, my initial workaround was just to
limit the number of rows retrieved from the Socrata API so I'll be able
to extract weather features for a lesser number of data.

However, there is a huge flaw in which I did not account that over
100,000 rows of data was only span across 3 days, which is clearly not
enough to train anything time-series related. So I streamlined the
script to have the ability to query based on a datetime range instead.
This does mean I have to reconsider how I want to collect weather data
though.