Added boundaries datasets and script to obtain it
Added a script to fetch the US Census TIGER boundaries by FIPS, which
then can be further broken down into custom user-defined boundaries.
It then reads the municipal issue dataset and maps each record's
latitude and longitude to respective boundaries.

As we'll be analysing and forecasting based on SG's planning zones,
looking into the details and patterns with more granularity could
provide more insightful training trends.

However, it is quite challenging to obtain the right level of
granularity due to the sheer number of GEOIDs within the tracts.
So perhaps, it might be more advisable to look into retrieving more
datasets from other cities? Unless a user-defined boundary is able to
encompass it into a good number of boundaries.