Updated weather script and removed datasets
As mentioned in a previous commit, by updating the municipal data
retrieval logic, we need to consider how we're going to fetch millions
of weather data while dealing with the 10k daily limit.

In any case, the original datasets were cleared for clarity, and I have
added threading (multi-processing) into the script, so we don't have to
wait so long for to go through each API call 1 by 1.

For now, I have updated and pushed this new script, so we all can use it
to query for weather data, which is 40k daily limit, not a lot but
definitely better than just 10k. Nonetheless, an alternative approach is
to group the latitudes and longitudes into cluster to reduce the sample
size, which hopefully doesn't dilute the weather data too much.