Added custom cluster dataset, with script
Added a script that utilises unsupervised learning, specificaly KMeans,
DBSCAN and HDBSCAN to cluster the latitude and longitudes of the
municipal reports by geographical proximity.

Apart from assigning the cluster number to the respective municipal
report of a given latitude and longtude, there will be another output file
given, which will contain the average of all the  latitude and longtitude
of points belonging to a specific cluster.

While I'm not sure how useful this is for analysis, this was done to
flatten the municipal dataset by assigning them into larger groups based
on proximity to address the issue of the 10k daily Weather API limit.